services/fastapi-service/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
langgraph==0.0.40
langchain==0.1.0
langchain-community==0.0.13
qdrant-client==1.7.1
minio==7.2.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
pyyaml==6.0.1
httpx==0.25.2
openai==1.3.0
google-generativeai==0.3.2
motor==3.3.2
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1